+++ internal/controllers/utils/utils.go
@@ -87,7 +87,7 @@ func IsIngressClassEmpty(obj client.Object) bool {
 }
 
 // CRDExists returns false if CRD does not exist.
-func CRDExists(client client.Client, gvr schema.GroupVersionResource) bool {
-	_, err := client.RESTMapper().KindFor(gvr)
+func CRDExists(restMapper meta.RESTMapper, gvr schema.GroupVersionResource) bool {
+	_, err := restMapper.KindFor(gvr)
 	return !meta.IsNoMatchError(err)
 }