+++ CHANGELOG.md
@@ -65,6 +65,12 @@ Adding a new version? You'll need three changes:
 
 > Release date: TBD
 
+### Added
+
+- Added `--cache-sync-timeout` flag allowing to change the default controllers' 
+  cache synchronisation timeout. 
+  [#3013](https://github.com/Kong/kubernetes-ingress-controller/pull/3013)
+
 ### Fixed
 
 - The controller now logs an error for and skips multi-Service rules that have
@@ -85,6 +91,10 @@ Adding a new version? You'll need three changes:
   makes the default the same target released as the standard
   `kong/kubernetes-ingress-controller:X.Y.Z` tags in the official repo.
   [#3043](https://github.com/Kong/kubernetes-ingress-controller/pull/3043)
+- The controller will no longer crash in case of missing CRDs installation.
+  Instead, an explicit message will be logged, informing that a given resource
+  controller has been disabled.
+  [#3013](https://github.com/Kong/kubernetes-ingress-controller/pull/3013)
 
 ## [2.7.0]